Move pandas dispatch registers to their own file
